+++ bio-seq/src/codec.rs
@@ -137,7 +137,7 @@ pub use bio_seq_derive::Codec;
 /// The intended representation is an `Enum`, transparently represented as a `u8`.
 pub trait Codec: fmt::Debug + Copy + Clone + PartialEq + Hash + Eq {
     /// The number of bits used to encode the symbols. e.g. `Dna::BITS` = 2, `Iupac::BITS` = 4.
-    const BITS: u8;
+    const BITS: usize;
 
     /// Convert raw bits of binary encoding into enum item. Binary values
     /// that don't match an enum member's discriminant will result in panic or random enum
