+++ bio-seq/src/codec/text.rs
@@ -10,7 +10,7 @@ use crate::error::ParseBioError;
 pub struct Dna(u8);
 
 impl Codec for Dna {
-    const BITS: u8 = 8;
+    const BITS: usize = 8;
 
     fn unsafe_from_bits(b: u8) -> Self {
         Self(b)
